修改了代码问题

+++ service/business/src/main/java/cn/fintecher/pangolin/business/web/AccTelPoolController.java
@@ -575,12 +575,12 @@ public class AccTelPoolController extends BaseController {
      */
     @PostMapping("/leaveTelCase")
     @ApiOperation(value = "电催案件留案操作", notes = "电催案件留案操作")
-    public ResponseEntity<Long> leaveTelCase(@RequestBody LeaveCaseParams leaveCaseParams,
+    public ResponseEntity<Integer> leaveTelCase(@RequestBody LeaveCaseParams leaveCaseParams,
                                              @RequestHeader(value = "X-UserToken") String token) {
         log.debug("REST request to leave case");
         try {
             User tokenUser = getUserByToken(token);
-            Long caseNum = caseInfoService.leaveCase(leaveCaseParams, tokenUser);
+            Integer caseNum = caseInfoService.leaveCase(leaveCaseParams, tokenUser);
             return ResponseEntity.ok().headers(HeaderUtil.createAlert("留案成功", ENTITY_CASEINFO)).body(caseNum);
         } catch (Exception e) {
             log.error(e.getMessage(), e);
